Fixes done

+++ model.py
@@ -212,7 +212,7 @@ class RMSNorm(nn.Module):
         x = x.float()
         variance = x.pow(2).mean(-1, keepdim=True)
         x = x * torch.rsqrt(variance + self.eps)
-        return (self.weight * x.to(dtype))
+        return self.weight.to(dtype) * x.to(dtype)
 
 
 # ---------------------------------------------------------------------------
